deploy_temp/requirements.txt
# Smart Brewery IoT Server Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database - Azure SQL
sqlalchemy==2.0.25
pymssql==2.2.11
pyodbc==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP
python-dotenv==1.0.0
python-multipart==0.0.6
